requirements.txt
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
typing_extensions  == 4.6.3
drf-yasg==1.21.6
